Implement SOURCE_DATE_EPOCH
If SOURCE_DATE_EPOCH is defined in the environment, clamp all timestamps
in the image to that maximum value.  This strips out uninteresting
differences in timestamps during the build process, making it easier to
reproduce a byte-for-byte identical image.

Note that with new versions of GNU tar, --clamp-mtime can do the same
thing, but this was not available in Debian jessie.

See https://wiki.debian.org/ReproducibleBuilds/TimestampsInTarball